pyproject.toml
[project]
name = "curius-obsidian"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx",
    "groq",
    "python-dotenv",
]
